+++ ConsentViewController/Classes/SourcePointClient/SourcepointClientCoordinator.swift
@@ -18,7 +18,6 @@ struct MessageToDisplay {
     let metadata: MessageMetaData
     let url: URL
     let type: SPCampaignType
-    let childPmId: String?
 }
 
 extension MessageToDisplay {
@@ -34,8 +33,6 @@ extension MessageToDisplay {
         self.metadata = metadata
         self.url = url
         self.type = campaign.type
-        // TODO: add childPMId if possible
-        self.childPmId = nil
     }
 }
 
@@ -44,8 +41,9 @@ protocol SPClientCoordinator {
     var deviceManager: SPDeviceManager { get set }
     var userData: SPUserData { get }
     var language: SPMessageLanguage { get set }
+    var spClient: SourcePointProtocol { get }
 
-    func loadMessages(forAuthId: String?, _ handler: @escaping MessagesAndConsentsHandler)
+    func loadMessages(forAuthId: String?, pubData: SPPublisherData?, _ handler: @escaping MessagesAndConsentsHandler)
     func reportAction(_ action: SPAction, handler: @escaping (Result<SPUserData, SPError>) -> Void)
     func reportIdfaStatus(status: SPIDFAStatus, osVersion: String)
     func logErrorMetrics(_ error: SPError)
@@ -135,7 +133,6 @@ class SourcepointClientCoordinator: SPClientCoordinator {
     var language: SPMessageLanguage
     var idfaStatus: SPIDFAStatus { SPIDFAStatus.current() }
     let campaigns: SPCampaigns
-    var pubData: SPPublisherData
 
     var deviceManager: SPDeviceManager
     let spClient: SourcePointProtocol
@@ -190,43 +187,6 @@ class SourcepointClientCoordinator: SPClientCoordinator {
         )
     }
 
-    var ccpaPvDataBodyFromState: PvDataRequestBody {
-        var ccpa: PvDataRequestBody.CCPA?
-        if let stateCCPA = state.ccpa {
-            ccpa = .init(
-                applies: stateCCPA.applies,
-                uuid: stateCCPA.uuid,
-                accountId: accountId,
-                siteId: propertyId,
-                consentStatus: stateCCPA.consentStatus,
-                pubData: pubData,
-                messageId: stateCCPA.lastMessage?.id,
-                sampleRate: state.ccpaMetaData?.sampleRate
-            )
-        }
-        return .init(ccpa: ccpa)
-    }
-    var gdprPvDataBodyFromState: PvDataRequestBody {
-        var gdpr: PvDataRequestBody.GDPR?
-        if let stateGDPR = state.gdpr {
-            gdpr = PvDataRequestBody.GDPR(
-                applies: stateGDPR.applies,
-                uuid: stateGDPR.uuid,
-                accountId: accountId,
-                siteId: propertyId,
-                consentStatus: stateGDPR.consentStatus,
-                pubData: pubData,
-                sampleRate: state.gdprMetaData?.sampleRate,
-                euconsent: stateGDPR.euconsent,
-                msgId: stateGDPR.lastMessage?.id,
-                categoryId: stateGDPR.lastMessage?.categoryId,
-                subCategoryId: stateGDPR.lastMessage?.subCategoryId,
-                prtnUUID: stateGDPR.lastMessage?.partitionUUID
-            )
-        }
-        return .init(gdpr: gdpr)
-    }
-
     var messagesParamsFromState: MessagesRequest {
         .init(
             body: .init(
@@ -279,7 +239,6 @@ class SourcepointClientCoordinator: SPClientCoordinator {
         propertyId: Int,
         language: SPMessageLanguage = .BrowserDefault,
         campaigns: SPCampaigns,
-        pubData: SPPublisherData = SPPublisherData(),
         storage: SPLocalStorage = SPUserDefaults(),
         spClient: SourcePointProtocol? = nil,
         deviceManager: SPDeviceManager = SPDevice.standard
@@ -289,7 +248,6 @@ class SourcepointClientCoordinator: SPClientCoordinator {
         self.propertyName = propertyName
         self.language = language
         self.campaigns = campaigns
-        self.pubData = pubData
         self.storage = storage
         self.spClient = spClient ?? SourcePointClient(
             accountId: accountId,
@@ -306,17 +264,57 @@ class SourcepointClientCoordinator: SPClientCoordinator {
     static func setupState(from localStorage: SPLocalStorage, campaigns localCampaigns: SPCampaigns) -> State {
         var localState = localStorage.spState ?? .init()
         if localCampaigns.gdpr != nil, localState.gdpr == nil {
-            localState.gdpr = .empty()
+            localState.gdpr = localStorage.userData.gdpr?.consents ?? .empty()
+            localState.gdpr?.applies = localStorage.userData.gdpr?.applies ?? false
             localState.gdprMetaData = .init()
         }
         if localCampaigns.ccpa != nil, localState.ccpa == nil {
-            localState.ccpa = .empty()
+            localState.ccpa = localStorage.userData.ccpa?.consents ?? .empty()
+            localState.ccpa?.applies = localStorage.userData.ccpa?.applies ?? false
             localState.ccpaMetaData = .init()
         }
         if localCampaigns.ios14 != nil, localState.ios14 == nil {
             localState.ios14 = .init()
         }
         return localState
+    }
+
+    func ccpaPvDataBody(from state: State, pubData: SPPublisherData?) -> PvDataRequestBody {
+        var ccpa: PvDataRequestBody.CCPA?
+        if let stateCCPA = state.ccpa {
+            ccpa = .init(
+                applies: stateCCPA.applies,
+                uuid: stateCCPA.uuid,
+                accountId: accountId,
+                siteId: propertyId,
+                consentStatus: stateCCPA.consentStatus,
+                pubData: pubData,
+                messageId: stateCCPA.lastMessage?.id,
+                sampleRate: state.ccpaMetaData?.sampleRate
+            )
+        }
+        return .init(ccpa: ccpa)
+    }
+
+    func gdprPvDataBody(from state: State, pubData: SPPublisherData?) -> PvDataRequestBody {
+        var gdpr: PvDataRequestBody.GDPR?
+        if let stateGDPR = state.gdpr {
+            gdpr = PvDataRequestBody.GDPR(
+                applies: stateGDPR.applies,
+                uuid: stateGDPR.uuid,
+                accountId: accountId,
+                siteId: propertyId,
+                consentStatus: stateGDPR.consentStatus,
+                pubData: pubData,
+                sampleRate: state.gdprMetaData?.sampleRate,
+                euconsent: stateGDPR.euconsent,
+                msgId: stateGDPR.lastMessage?.id,
+                categoryId: stateGDPR.lastMessage?.categoryId,
+                subCategoryId: stateGDPR.lastMessage?.subCategoryId,
+                prtnUUID: stateGDPR.lastMessage?.partitionUUID
+            )
+        }
+        return .init(gdpr: gdpr)
     }
 
     /// Resets state if the authId has changed, except if the stored auth id was empty.
@@ -340,7 +338,7 @@ class SourcepointClientCoordinator: SPClientCoordinator {
         storage.spState = state
     }
 
-    func loadMessages(forAuthId authId: String?, _ handler: @escaping MessagesAndConsentsHandler) {
+    func loadMessages(forAuthId authId: String?, pubData: SPPublisherData?, _ handler: @escaping MessagesAndConsentsHandler) {
         state = Self.setupState(from: storage, campaigns: campaigns)
         storage.spState = state
 
@@ -350,7 +348,7 @@ class SourcepointClientCoordinator: SPClientCoordinator {
             self.consentStatus {
                 self.state.udpateGDPRStatus()
                 self.messages { messagesResponse in
-                    self.pvData {
+                    self.pvData(pubData: pubData) {
                         handler(messagesResponse)
                     }
                 }
@@ -563,18 +561,18 @@ class SourcepointClientCoordinator: SPClientCoordinator {
         }
     }
 
-    func pvData(_ handler: @escaping () -> Void) {
+    func pvData(pubData: SPPublisherData?, _ handler: @escaping () -> Void) {
         let pvDataGroup = DispatchGroup()
         if let gdprMetadata = state.gdprMetaData {
             pvDataGroup.enter()
-            let sampled = sampleAndPvData(gdprMetadata, body: gdprPvDataBodyFromState) {
+            let sampled = sampleAndPvData(gdprMetadata, body: gdprPvDataBody(from: state, pubData: pubData)) {
                 pvDataGroup.leave()
             }
             state.gdprMetaData?.wasSampled = sampled
         }
         if let ccpaMetadata = state.ccpaMetaData {
             pvDataGroup.enter()
-            let sampled = sampleAndPvData(ccpaMetadata, body: ccpaPvDataBodyFromState) {
+            let sampled = sampleAndPvData(ccpaMetadata, body: ccpaPvDataBody(from: state, pubData: pubData)) {
                 pvDataGroup.leave()
             }
             state.ccpaMetaData?.wasSampled = sampled
@@ -640,7 +638,7 @@ class SourcepointClientCoordinator: SPClientCoordinator {
                 messageId: String(state.gdpr?.lastMessage?.id ?? 0),
                 consentAllRef: postPayloadFromGetCall?.consentAllRef,
                 vendorListId: postPayloadFromGetCall?.vendorListId,
-                pubData: action.publisherData,
+                pubData: action.encodablePubData,
                 pmSaveAndExitVariables: action.pmPayload,
                 sendPVData: state.gdprMetaData?.wasSampled ?? false,
                 propertyId: propertyId,
@@ -661,7 +659,7 @@ class SourcepointClientCoordinator: SPClientCoordinator {
                 authId: authId,
                 uuid: state.ccpa?.uuid,
                 messageId: String(state.ccpa?.lastMessage?.id ?? 0),
-                pubData: action.publisherData,
+                pubData: action.encodablePubData,
                 pmSaveAndExitVariables: action.pmPayload,
                 sendPVData: state.ccpaMetaData?.wasSampled ?? false,
                 propertyId: propertyId,